Add/Remove mocking test
