+++ addremove.js
@@ -1,5 +1,3 @@
-const setStored = () => localStorage.setItem('newItems', JSON.stringify(items));
-
 const addCrud = () => {
   const addEnter = document.querySelector('.add-enter');
   addEnter.addEventListener('keyup', ({ key }) => {
@@ -14,15 +12,6 @@ const addCrud = () => {
   });
 };
 
-class AddItem {
-  constructor(description, index) {
-    this.description = description;
-    this.completed = false;
-    this.index = index;
-    this.id = `item-${this.index}`;
-  }
-}
-
 const removeCrud = (trashIcon) => {
   trashIcon.addEventListener('click', () => {
     const filtered = items.filter((item) => item.id !== trashIcon.id);
@@ -34,41 +23,5 @@ const removeCrud = (trashIcon) => {
     displayList();
   });
 };
-
-function displayList() {
-  const list = getStored();
-  if (list !== null && list !== undefined) {
-    items = list;
-  }
-  const oldList = document.querySelectorAll('.container');
-  [...oldList].forEach((e) => e.remove());
-  items.forEach((item, i) => {
-    items[i].index = i;
-    const container = document.createElement('div');
-    const input = document.createElement('input');
-    const label = document.createElement('label');
-    const moreIcon = document.createElement('img');
-    moreIcon.src = more;
-    const trashIcon = document.createElement('img');
-    trashIcon.src = trash;
-    container.appendChild(input);
-    container.appendChild(label);
-    container.appendChild(moreIcon);
-    container.appendChild(trashIcon);
-    container.classList.add('container');
-    input.classList.add('checkbox');
-    label.classList.add('edit');
-    trashIcon.id = item.id;
-    input.setAttribute('type', 'checkbox');
-    label.setAttribute('contenteditable', 'true');
-    const textNode = document.createTextNode(item.description);
-    label.appendChild(textNode);
-    checkList.appendChild(container);
-    check(input, item, label, setStored, trashIcon, moreIcon);
-    editCrud();
-    removeCrud(trashIcon);
-    getStored();
-  });
-}
 
 export { addCrud, removeCrud }
